requirements.txt
flask
tensorflow-cpu
opencv-python-headless
numpy
werkzeug